+++ custom_components/supernotify/media_grab.py
@@ -2,6 +2,7 @@ import asyncio
 import io
 import logging
 import time
+from enum import StrEnum, auto
 from http import HTTPStatus
 from io import BytesIO
 from pathlib import Path
@@ -21,6 +22,7 @@ from custom_components.supernotify import (
     ATTR_MEDIA_CAMERA_ENTITY_ID,
     ATTR_MEDIA_CAMERA_PTZ_PRESET,
     ATTR_MEDIA_SNAPSHOT_URL,
+    ATTR_PNG_OPTS,
     CONF_ALT_CAMERA,
     CONF_CAMERA,
     CONF_DEVICE_TRACKER,
@@ -28,7 +30,9 @@ from custom_components.supernotify import (
     CONF_PTZ_DELAY,
     CONF_PTZ_METHOD,
     CONF_PTZ_PRESET_DEFAULT,
+    MEDIA_OPTION_REPROCESS,
     OPTION_JPEG,
+    OPTION_PNG,
     PTZ_METHOD_FRIGATE,
     PTZ_METHOD_ONVIF,
 )
@@ -41,6 +45,12 @@ if TYPE_CHECKING:
 _LOGGER = logging.getLogger(__name__)
 
 
+class ReprocessOption(StrEnum):
+    ALWAYS = auto()
+    NEVER = auto()
+    PRESERVE = auto()
+
+
 async def snapshot_from_url(
     hass: HomeAssistant | None,
     snapshot_url: str,
@@ -48,7 +58,9 @@ async def snapshot_from_url(
     media_path: Path,
     hass_base_url: str | None,
     remote_timeout: int = 15,
+    reprocess: ReprocessOption = ReprocessOption.ALWAYS,
     jpeg_opts: dict[str, Any] | None = None,
+    png_opts: dict[str, Any] | None = None,
 ) -> Path | None:
     hass_base_url = hass_base_url or ""
     if not hass:
@@ -66,37 +78,18 @@ async def snapshot_from_url(
         if r.status != HTTPStatus.OK:
             _LOGGER.warning("SUPERNOTIFY Unable to retrieve %s: %s", image_url, r.status)
         else:
-            if r.content_type in ("image/jpeg", "image/jpg"):
-                media_ext = "jpg"
-                image_format = "JPEG"
-            elif r.content_type == "image/png":
-                media_ext = "png"
-                image_format = "PNG"
-            elif r.content_type == "image/gif":
-                media_ext = "gif"
-                image_format = "GIF"
-            else:
-                _LOGGER.info("SUPERNOTIFY Unexpected MIME type %s from snap of %s", r.content_type, image_url)
-                media_ext = "img"
-                image_format = None
+            bitmap: bytes | None = await r.content.read()
+            image_path: anyio.Path | None = await write_image_from_bitmap(
+                bitmap, media_path, notification_id, reprocess=reprocess, jpeg_opts=jpeg_opts, png_opts=png_opts
+            )
+            if image_path:
+                _LOGGER.debug("SUPERNOTIFY Fetched image from %s to %s", image_url, image_path)
+                return Path(image_path)
 
-            # TODO: configure image rewrite
-            image_path: Path = Path(media_dir) / f"{notification_id}.{media_ext}"
-            image: Image.Image = Image.open(io.BytesIO(await r.content.read()))
-            # rewrite to remove metadata, incl custom CCTV comments that confusie python MIMEImage
-            clean_image: Image.Image = Image.new(image.mode, image.size)
-            clean_image.putdata(image.getdata())
-            buffer = BytesIO()
-            img_args = {}
-            if image_format == "JPEG" and jpeg_opts:
-                img_args.update(jpeg_opts)
-            clean_image.save(buffer, image_format, **img_args)
-            async with aiofiles.open(image_path, "wb") as file:
-                await file.write(buffer.getbuffer())
-            _LOGGER.debug("SUPERNOTIFY Fetched image from %s to %s", image_url, image_path)
-            return image_path
+        _LOGGER.warning("SUPERNOTIFY Failed to snap image from %s", snapshot_url)
     except Exception as e:
         _LOGGER.error("SUPERNOTIFY Image snap fail: %s", e)
+
     return None
 
 
@@ -129,12 +122,14 @@ async def move_camera_to_ptz_preset(
         _LOGGER.warning("SUPERNOTIFY Unable to move %s to ptz preset %s: %s", camera_entity_id, preset, e)
 
 
-async def snap_image(
+async def snap_image_entity(
     context: Context,
     entity_id: str,
     media_path: Path,
     notification_id: str,
+    reprocess: ReprocessOption = ReprocessOption.ALWAYS,
     jpeg_opts: dict[str, Any] | None = None,
+    png_opts: dict[str, Any] | None = None,
 ) -> Path | None:
     """Use for any image, including MQTT Image"""
     image_path: anyio.Path | None = None
@@ -145,44 +140,30 @@ async def snap_image(
             image_entity = context.hass_api._hass.data["image"].get_entity(entity_id)
         if image_entity:
             bitmap: bytes | None = await image_entity.async_image()
-            if bitmap is None:
-                _LOGGER.warning("SUPERNOTIFY Empty bitmap from image entity %s", entity_id)
-            else:
-                image: Image.Image = Image.open(io.BytesIO(bitmap))
-                media_dir: anyio.Path = anyio.Path(media_path) / "image"
-                await media_dir.mkdir(parents=True, exist_ok=True)
-
-                media_ext: str = image.format.lower() if image.format else "img"
-                timed: str = str(time.time()).replace(".", "_")
-                image_path = anyio.Path(media_dir) / f"{notification_id}_{timed}.{media_ext}"
-                buffer = BytesIO()
-                img_args = {}
-                if media_ext in ("jpg", "jpeg") and jpeg_opts:
-                    img_args.update(jpeg_opts)
-                image.save(buffer, image.format, **img_args)
-                async with aiofiles.open(await image_path.resolve(), "wb") as file:
-                    await file.write(buffer.getbuffer())
-        else:
-            _LOGGER.warning("SUPERNOTIFY Unable to find image entity %s", entity_id)
+            image_path = await write_image_from_bitmap(
+                bitmap, media_path, notification_id, reprocess=reprocess, jpeg_opts=jpeg_opts, png_opts=png_opts
+            )
     except Exception as e:
         _LOGGER.warning("SUPERNOTIFY Unable to snap image %s: %s", entity_id, e)
-        return None
-    return Path(await image_path.resolve()) if image_path else None
+    if image_path is None:
+        _LOGGER.warning("SUPERNOTIFY Unable to save from image entity %s", entity_id)
+    return Path(image_path) if image_path else None
 
 
 async def snap_camera(
     hass: HomeAssistant,
     camera_entity_id: str,
+    notification_id: str,
     media_path: Path,
     max_camera_wait: int = 20,
+    reprocess: ReprocessOption = ReprocessOption.ALWAYS,
     jpeg_opts: dict[str, Any] | None = None,
+    png_opts: dict[str, Any] | None = None,
 ) -> Path | None:
     image_path: Path | None = None
     if not camera_entity_id:
         _LOGGER.warning("SUPERNOTIFY Empty camera entity id for snap")
         return image_path
-    if jpeg_opts:
-        _LOGGER.warning("jpeg_opts not yet supported by snap_camera")
 
     try:
         media_dir: anyio.Path = anyio.Path(media_path) / "camera"
@@ -198,6 +179,17 @@ async def snap_camera(
         while time.time() < cutoff_time and not image_path.exists():
             _LOGGER.info("Image file not available yet at %s, pausing", image_path)
             await asyncio.sleep(1)
+
+        if reprocess != ReprocessOption.NEVER:
+            async with await anyio.Path(image_path).open("rb") as f:
+                bitmap: bytes | None = await f.read()
+                async_path: anyio.Path | None = await write_image_from_bitmap(
+                    bitmap, media_path, notification_id, reprocess=reprocess, jpeg_opts=jpeg_opts, png_opts=png_opts
+                )
+                if async_path:
+                    image_path = Path(async_path)
+                else:
+                    _LOGGER.warning("SUPERNOTIFY Unable to reprocess camera image")
 
     except Exception as e:
         _LOGGER.warning("Failed to snap avail camera %s to %s: %s", camera_entity_id, image_path, e)
@@ -262,6 +254,17 @@ async def grab_image(notification: "Notification", delivery_name: str, context: 
     camera_entity_id = notification.media.get(ATTR_MEDIA_CAMERA_ENTITY_ID)
     delivery_config = notification.delivery_data(delivery_name)
     jpeg_opts = notification.media.get(ATTR_JPEG_OPTS, delivery_config.get(CONF_OPTIONS, {}).get(OPTION_JPEG))
+    png_opts = notification.media.get(ATTR_PNG_OPTS, delivery_config.get(CONF_OPTIONS, {}).get(OPTION_PNG))
+    reprocess_option = (
+        notification.media.get(MEDIA_OPTION_REPROCESS, delivery_config.get(CONF_OPTIONS, {}).get(MEDIA_OPTION_REPROCESS))
+        or "always"
+    )
+
+    reprocess: ReprocessOption = ReprocessOption.ALWAYS
+    try:
+        reprocess = ReprocessOption(reprocess_option)
+    except Exception:
+        _LOGGER.warning("SUPERNOTIFY Invalid reprocess option: %s", reprocess_option)
 
     if not snapshot_url and not camera_entity_id:
         return None
@@ -276,13 +279,23 @@ async def grab_image(notification: "Notification", delivery_name: str, context: 
             notification.id,
             context.media_path,
             context.hass_api.internal_url,
-            jpeg_opts,
+            reprocess=reprocess,
+            jpeg_opts=jpeg_opts,
+            png_opts=png_opts,
         )
     elif camera_entity_id and camera_entity_id.startswith("image.") and context.hass_api._hass and context.media_path:
-        image_path = await snap_image(context, camera_entity_id, context.media_path, notification.id, jpeg_opts)
+        image_path = await snap_image_entity(
+            context,
+            camera_entity_id,
+            context.media_path,
+            notification.id,
+            reprocess=reprocess,
+            jpeg_opts=jpeg_opts,
+            png_opts=png_opts,
+        )
     elif camera_entity_id:
         if not context.hass_api._hass or not context.media_path:
-            _LOGGER.warning("SUPERNOTIFY No homeassistant ref or media path for camera %s", camera_entity_id)
+            _LOGGER.warning("SUPERNOTIFY No HA ref or media path for camera %s", camera_entity_id)
             return None
         active_camera_entity_id = select_avail_camera(context.hass_api._hass, context.cameras, camera_entity_id)
         if active_camera_entity_id:
@@ -308,9 +321,12 @@ async def grab_image(notification: "Notification", delivery_name: str, context: 
             image_path = await snap_camera(
                 context.hass_api._hass,
                 active_camera_entity_id,
+                notification.id,
+                reprocess=reprocess,
                 media_path=context.media_path,
                 max_camera_wait=15,
                 jpeg_opts=jpeg_opts,
+                png_opts=png_opts,
             )
             if camera_ptz_preset and camera_ptz_preset_default:
                 await move_camera_to_ptz_preset(
@@ -322,4 +338,56 @@ async def grab_image(notification: "Notification", delivery_name: str, context: 
         return None
     # TODO: replace poking inside notification
     notification.snapshot_image_path = image_path
+    return image_path
+
+
+async def write_image_from_bitmap(
+    bitmap: bytes | None,
+    media_path: Path,
+    notification_id: str,
+    reprocess: ReprocessOption = ReprocessOption.ALWAYS,
+    output_format: str | None = None,
+    jpeg_opts: dict[str, Any] | None = None,
+    png_opts: dict[str, Any] | None = None,
+) -> anyio.Path | None:
+    image_path: anyio.Path | None = None
+    if bitmap is None:
+        _LOGGER.debug("SUPERNOTIFY Empty bitmap for image")
+        return None
+    try:
+        media_dir: anyio.Path = anyio.Path(media_path) / "image"
+        if not await media_dir.exists():
+            await media_dir.mkdir(parents=True, exist_ok=True)
+
+        image: Image.Image = Image.open(io.BytesIO(bitmap))
+        input_format = image.format.lower() if image.format else "img"
+        if reprocess == ReprocessOption.ALWAYS:
+            # rewrite to remove metadata, incl custom CCTV comments that confusie python MIMEImage
+            clean_image: Image.Image = Image.new(image.mode, image.size)
+            clean_image.putdata(image.getdata())
+            image = clean_image
+
+        buffer = BytesIO()
+        img_args = {}
+        if reprocess in (ReprocessOption.ALWAYS, ReprocessOption.PRESERVE):
+            if input_format in ("jpg", "jpeg") and jpeg_opts:
+                img_args.update(jpeg_opts)
+            elif input_format == "png" and png_opts:
+                img_args.update(png_opts)
+
+        output_format = output_format or input_format
+        image.save(buffer, output_format, **img_args)
+
+        media_ext: str = output_format if output_format else "img"
+        timed: str = str(time.time()).replace(".", "_")
+        image_path = anyio.Path(media_dir) / f"{notification_id}_{timed}.{media_ext}"
+        image_path = await image_path.resolve()
+        async with aiofiles.open(image_path, "wb") as file:
+            await file.write(buffer.getbuffer())
+    except TypeError:
+        # probably a jpeg or png option
+        _LOGGER.exception("SUPERNOTIFY Image snap fail")
+    except Exception:
+        _LOGGER.exception("SUPERNOTIFY Failure saving %s bitmap", input_format)
+        image_path = None
     return image_path